+++ src/client.rs
@@ -177,6 +177,7 @@ impl Fetch for MockClient {
 #[cfg(test)]
 mod tests {
     use super::*;
+    use std::collections::HashSet;
     use tokio::time::Instant;
 
     #[tokio::test]
@@ -189,11 +190,13 @@ mod tests {
             "https://en.wikipedia.org/wiki/Applicative_functor",
             now,
             None,
+            HashSet::new(),
         );
         let bookmark2 = TargetBookmark::new(
             "https://en.wikipedia.org/wiki/Monad_(functional_programming)",
             now,
             None,
+            HashSet::new(),
         );
 
         let start_instant = Instant::now();
@@ -217,11 +220,13 @@ mod tests {
             "https://en.wikipedia.org/wiki/Applicative_functor",
             now,
             None,
+            HashSet::new(),
         );
         let bookmark2 = TargetBookmark::new(
             "https://en.wikipedia.org/wiki/Monad_(functional_programming)",
             now,
             None,
+            HashSet::new(),
         );
 
         let last_fetched = throttler.last_fetched(&bookmark1, now).unwrap();